Fixing Relabel and RelabelObject plugin.
The Relabel plugin was renamed as RelabelObject, but was not removed
from the repository. Now, we take it out. The RelabelObjects plugin
is now using the progress reporting mechanism and the Qt dialog for
requesting numeric parameters, in particular the minimum object size
that should be considered when relabeling the connected components.